requirements.txt
discord.py
pyyaml